refactor extend typescript support to dialog linkTd, property and template
