+++ src/components/Dialogs/DialogTemplate.tsx
@@ -10,9 +10,20 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR W3C-20150513
  ********************************************************************************/
-import React from "react";
+import React, { ReactNode, useEffect } from "react";
 
-export const DialogTemplate = (props) => {
+interface DialogTemplateProps {
+  title?: string;
+  description?: string;
+  children?: ReactNode;
+  cancelText?: string;
+  onCancel?: () => void;
+  submitText?: string;
+  onSubmit?: () => void;
+  hasSubmit?: boolean;
+}
+
+const DialogTemplate: React.FC<DialogTemplateProps> = (props) => {
   const title = props.title ?? "Default Title";
   const description =
     props.description ?? "Default description, please override.";
@@ -24,17 +35,18 @@ export const DialogTemplate = (props) => {
   const submitText = props.submitText ?? "Submit";
   const onSubmit = props.onSubmit ?? (() => {});
   const hasSubmit = props.hasSubmit ?? true;
+
   let keysDown = {};
-  window.onkeydown = function (e) {
+  window.onkeydown = function (e: KeyboardEvent) {
     keysDown[e.key] = true;
     if (keysDown["Control"] && keysDown["Enter"]) {
-      document.getElementById("submitButton").click();
+      document.getElementById("submitButton")?.click();
     } else if (keysDown["Escape"]) {
-      document.getElementById("cancelButton").click();
+      document.getElementById("cancelButton")?.click();
     }
   };
 
-  window.onkeyup = function (e) {
+  window.onkeyup = function (e: KeyboardEvent) {
     keysDown[e.key] = false;
   };
 
@@ -50,7 +62,7 @@ export const DialogTemplate = (props) => {
           <div className="flex justify-end p-2 pt-4">
             <button
               id="cancelButton"
-              className="mr-1 rounded-md bg-gray-500 p-2 text-white"
+              className="mr-1 rounded-md bg-blue-500 p-2 text-white hover:bg-blue-600"
               onClick={() => {
                 onCancel();
               }}
@@ -60,7 +72,7 @@ export const DialogTemplate = (props) => {
             {hasSubmit && (
               <button
                 id="submitButton"
-                className="flex rounded-md bg-blue-500 p-2 text-white"
+                className="flex rounded-md bg-blue-500 p-2 text-white hover:bg-blue-600"
                 onClick={() => onSubmit()}
               >
                 {submitText}
@@ -72,3 +84,5 @@ export const DialogTemplate = (props) => {
     </>
   );
 };
+
+export default DialogTemplate;
